move some code in src/ to more logical locations
