+++ src/Makefile
@@ -36,10 +36,10 @@ FLAGS += -DJL_BUILD_UNAME='"$(OS)"'
 endif
 
 SRCS := \
-	jltypes gf typemap ast builtins module interpreter \
-	alloc dlload sys init task array dump toplevel jl_uv \
+	jltypes gf typemap ast builtins module interpreter symbol \
+	dlload sys init task array dump toplevel jl_uv datatype \
 	simplevector APInt-C runtime_intrinsics runtime_ccall \
-	threadgroup threading stackwalk gc gc-debug gc-pages \
+	threadgroup threading stackwalk gc gc-debug gc-pages method \
 	jlapi signal-handling safepoint jloptions timing subtype
 
 ifeq ($(USEMSVC), 1)
